+++ handler_impl.go
@@ -6,6 +6,9 @@ import (
 	"fmt"
 	"io/ioutil"
 	goHttp "net/http"
+	"sort"
+	"strconv"
+	"strings"
 
 	"github.com/containerssh/log"
 )
@@ -28,8 +31,11 @@ func (s *serverResponse) SetBody(body interface{}) {
 }
 
 type handler struct {
-	requestHandler RequestHandler
-	logger         log.Logger
+	requestHandler            RequestHandler
+	logger                    log.Logger
+	defaultResponseMarshaller responseMarshaller
+	defaultResponseType       string
+	responseMarshallers       []responseMarshaller
 }
 
 var internalErrorResponse = serverResponse{
@@ -60,7 +66,11 @@ func (h *handler) ServeHTTP(goWriter goHttp.ResponseWriter, goRequest *goHttp.Re
 			response = internalErrorResponse
 		}
 	}
-	bytes, err := json.Marshal(response.body)
+	marshaller, responseType, statusCode := h.findMarshaller(goWriter, goRequest)
+	if statusCode != 200 {
+		goWriter.WriteHeader(statusCode)
+	}
+	bytes, err := marshaller.Marshal(response.body)
 	if err != nil {
 		h.logger.Error(log.Wrap(err, MServerEncodeFailed, "failed to marshal response %v", response))
 		response = internalErrorResponse
@@ -71,10 +81,51 @@ func (h *handler) ServeHTTP(goWriter goHttp.ResponseWriter, goRequest *goHttp.Re
 		}
 	}
 	goWriter.WriteHeader(int(response.statusCode))
-	goWriter.Header().Add("Content-Type", "application/json")
+	goWriter.Header().Add("Content-Type", responseType)
 	if _, err := goWriter.Write(bytes); err != nil {
 		h.logger.Debug(log.Wrap(err, MServerResponseWriteFailed, "Failed to write HTTP response"))
 	}
+}
+
+func (h *handler) findMarshaller(_ goHttp.ResponseWriter, request *goHttp.Request) (responseMarshaller, string, int) {
+	acceptHeader := request.Header.Get("Accept")
+	if acceptHeader == "" {
+		return h.defaultResponseMarshaller, h.defaultResponseType, 200
+	}
+
+	accepted := strings.Split(acceptHeader, ",")
+	acceptMap := make(map[string]float64, len(accepted))
+	acceptList := make([]string, len(accepted))
+	for i, accept := range accepted {
+		acceptParts := strings.SplitN(strings.TrimSpace(accept), ";", 2)
+		q := 1.0
+		if len(acceptParts) == 2 {
+			acceptParts2 := strings.SplitN(acceptParts[1], "=", 2)
+			if acceptParts2[0] == "q" && len(acceptParts2) == 2 {
+				var err error
+				q, err = strconv.ParseFloat(acceptParts2[1], 64)
+				if err != nil {
+					return nil, h.defaultResponseType, 400
+				}
+			} else {
+				return nil, h.defaultResponseType, 400
+			}
+		}
+		acceptMap[acceptParts[0]] = q
+		acceptList[i] = acceptParts[0]
+	}
+	sort.SliceStable(acceptList, func(i, j int) bool {
+		return acceptMap[acceptList[i]] > acceptMap[acceptList[j]]
+	})
+
+	for _, a := range acceptList {
+		for _, marshaller := range h.responseMarshallers {
+			if marshaller.SupportsMIME(a) {
+				return marshaller, a, 200
+			}
+		}
+	}
+	return nil, h.defaultResponseType, 406
 }
 
 type internalRequest struct {